% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_price.R
\name{total_price}
\alias{total_price}
\title{Calculate total price with Discount by currency}
\usage{
total_price(.data, .currency = "EUR")
}
\arguments{
\item{.data}{A data.frame/tibble of orders.}

\item{.currency}{A character string that depicts the desired currency.}
}
\description{
\code{total_price} This function calculates total price with Discount by currency.
}
